Expose delayed jobs as an API route
This patch exposes delayed jobs data as an API route, which makes it
accessible using a bearer token. Currently, a token is not accepted as
valid authentication, and an Unauthorized error is returned to the
client.

Test plan:

- create a user in the Site Admin account and grant it admin rights to
  Site Admin
- create a user in the Site Admin account (or any other account) but do
  not grant it admin rights to Site Admin
- create some delayed jobs, failed and otherwise
- test the following routes:
  -`GET /api/v1/jobs?only=jobs&flavor=[failed|future|current|waiting]`:
should return list of jobs for specified flavor in JSON
  - `POST /api/v1/jobs/batch_update` with body
    `update_action=destroy&job_ids[]=$id1&job_ids[]=$id2`: (replacing
$id1 and $id2 with real job IDs) should return a success message with
count of destroyed jobs. If you pass the IDs of failed jobs, they
currently won't be deleted (pending
[instructure/canvas-jobs#2](https://github.com/instructure/canvas-jobs/pulls/2)
  - `POST /api/v1/jobs/batch_update` with body
    `update_action=destroy&flavor=future`: should return a success
message with count of destroyed jobs

When testing the routes, using the second user created (without Site
Admin rights) should throw an Unauthorized error. Using a developer key
should throw an Invalid Token error.